+++ filament_scale_enhanced/static/js/filament_scale.js
