+++ filament_scale_enhanced/templates/filament_scale_settings.jinja2
